Fetch & render data from models.py page to page
POST data : data is fetched via from models.py file to forms.py file under the meta class from where it is imported to the views.py page under a function, thereby it is rendered on a html page.
FETCH DATA : data is fetched via a database query within the views.py function. Then it is rendered to the html page. Focus on the database query language within the django app